+++ app/Models/Sys/Vm.php
@@ -8,4 +8,37 @@ use Illuminate\Database\Eloquent\Model;
 class Vm extends Model
 {
     use HasFactory;
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = [
+        'vm_id',
+        'name',
+        'pri_host',
+        'pub_host',
+    ];
+
+    /**
+     * Make attributes available in the json response.
+     *
+     * @var array
+     */
+    protected $appends = [];
+
+    /**
+     * The model's default values for attributes.
+     *
+     * @var array
+     */
+    protected $attributes = [];
+
+    /**
+     * Indicates if the model should be timestamped.
+     *
+     * @var bool
+     */
+    public $timestamps = true;
 }